Fix bug preventing users from selecting midnight as a study time.
The time dropdowns were using "0" to mean that no time was selected and
"0000" to mean midnight. However, during validation the check to see if
the user didn't enter a value used ($value == "0") which, since the
double-equals operator tries to cast each side to an int if possible,
meant that midnight was interpreted as no value because "0000" == "0".
You couldn't select midnight.

Use the empty string "" as the default value instead, and use
triple-equals === to compare.