Prevent accidentally submitting the new group form with the add member button.
It can't be a separate form because we don't want to lose the rest of
the data when it is submitted, and i don't want to rely on the browser
submitting button values to determine which button was pressed. So add
some javascript to prevent submitting the form when the add user field
is empty.